widget_refactoring
