+++ jordan/pubspec.yaml
@@ -23,8 +23,7 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-
-
+        
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
